[build](be) make master compile on aarch64 Ubuntu 22.04 (clang 19)
- bitmap_value.h: suppress -Wshadow for BitmapDataType enum constants
  (they collide with protobuf-exported doris::BITMAP; clang >= 17 turns
  it into an -Werror failure)
- be/CMakeLists.txt: keep -Wshadow non-fatal on clang, and
  -Wno-deprecated-declarations for BE_TEST (libstdc++ >= 12 marks
  std::get_temporary_buffer deprecated, tripping existing tests)
- parquet_column_convert.h: std::powf -> std::pow (libstdc++ <cmath>
  does not declare powf in namespace std)
- function_string_misc.cpp: std::format -> fmt::format (libstdc++ 12
  has no <format>; fmt is the project convention)
- be_thread_stack_action.cpp: append_frame is only used on the x86_64
  unwind path; mark [[maybe_unused]] so aarch64 -Werror builds pass
- glibc-compatibility: add resolv_shim.c forwarding __res_nsearch to
  res_nsearch — glibc >= 2.34 exports the former only as a non-default
  compat version, breaking the link against the prebuilt krb5 archive

+++ be/src/format/parquet/parquet_column_convert.h
@@ -22,6 +22,7 @@
 #include <libdivide.h>
 
 #include <chrono>
+#include <cmath> // std::pow (used by the half-float decoding path)
 #include <limits>
 
 #include "common/cast_set.h"
@@ -593,7 +594,7 @@ public:
                 // half subnormal:
                 //    value = (-1)^sign * (mant / 2^10) * 2^(1 - bias)
                 // half bias = 15 → exponent = 1 - 15 = -14
-                float f = (static_cast<float>(mant) / 1024.0F) * std::powf(2.0F, -14.0F);
+                float f = (static_cast<float>(mant) / 1024.0F) * std::pow(2.0F, -14.0F);
                 return sign ? -f : f;
             }
         } else if (exp == 0x1F) {
